# shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/packs": {
      "get": {
        "operationId": "PackController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "status": {
                        "type": "string"
                      },
                      "unlocked": {
                        "type": "boolean"
                      },
                      "locked": {
                        "type": "boolean"
                      },
                      "freeCount": {
                        "type": "integer"
                      },
                      "packSize": {
                        "type": "integer"
                      },
                      "regensLeft": {
                        "type": "integer"
                      },
                      "stickers": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "index": {
                              "type": "integer"
                            },
                            "url": {
                              "type": "string"
                            }
                          },
                          "required": ["index", "url"]
                        }
                      }
                    },
                    "required": [
                      "id",
                      "createdAt",
                      "status",
                      "unlocked",
                      "locked",
                      "freeCount",
                      "packSize",
                      "regensLeft",
                      "stickers"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["packs"]
      },
      "post": {
        "operationId": "PackController_create",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "packId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": ["packId"]
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["packs"]
      }
    },
    "/packs/{packId}": {
      "get": {
        "operationId": "PackController_getOne",
        "parameters": [
          {
            "name": "packId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string"
                    },
                    "unlocked": {
                      "type": "boolean"
                    },
                    "locked": {
                      "type": "boolean"
                    },
                    "freeCount": {
                      "type": "integer"
                    },
                    "packSize": {
                      "type": "integer"
                    },
                    "regensLeft": {
                      "type": "integer"
                    },
                    "stickers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "index": {
                            "type": "integer"
                          },
                          "url": {
                            "type": "string"
                          }
                        },
                        "required": ["index", "url"]
                      }
                    },
                    "selfieUrl": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "unlocked",
                    "locked",
                    "freeCount",
                    "packSize",
                    "regensLeft",
                    "stickers"
                  ]
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["packs"]
      },
      "delete": {
        "operationId": "PackController_remove",
        "parameters": [
          {
            "name": "packId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pack deleted"
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["packs"]
      }
    },
...
    "/packs/{packId}/download": {
      "post": {
        "operationId": "PackController_download",
        "parameters": [
          {
            "name": "packId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "locked": {
                      "type": "boolean"
                    }
                  },
                  "required": ["locked"]
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["packs"]
      }
    },
    "/packs/{packId}/claim-free": {
      "post": {
        "operationId": "PackController_claimFree",
        "parameters": [
          {
            "name": "packId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "delivered": {
                      "type": "boolean"
                    },
                    "botUrl": {
                      "type": "string",
                      "format": "uri"
                    },
                    "alreadyClaimed": {
                      "type": "boolean",
                      "nullable": true
                    }
                  },
                  "required": ["delivered", "botUrl"]
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["packs"]
      }
    },
...
  },
  "info": {
    "title": "Stickup API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      }
    },
...
  }
}